Move model related methods when Model is included only.
