--- lib/trailblazer/rails/autoloading.rb
@@ -1,3 +1,0 @@
-Trailblazer::Operation.class_eval do
-  autoload :Responder,  "trailblazer/operation/responder"
-end
